Document Derivative (#366)
* Document Derivative

* Mention absolute
